refactor: change file colocation
