+++ js/src/forum/extenders/addFrontPage.js
