+++ js/src/forum/extenders/addStickyBadge.js
