+++ js/src/forum/index.ts
@@ -1,7 +1,7 @@
 import app from 'flarum/forum/app';
-import addFrontSort from './addFrontSort';
-import addFrontPage from './addFrontPage';
-import addStickyBadge from './addStickyBadge';
+import addFrontSort from './extenders/addFrontSort';
+import addFrontPage from './extenders/addFrontPage';
+import addStickyBadge from './extenders/addStickyBadge';
 
 app.initializers.add('fof/frontpage', () => {
   addFrontSort();
